day 11 aoc 2020: game of life in smooth C++ 🤩
the diff is so simple too!

We only change the behaviour of seat counting - exactly as required.
And we do that with just a simple `multiplier` var,
and then `break`ing early if the seat is not the floor
since you cannot see past other seats.

The required count changes from `4` to `5` - that's a single digit change.

And that's it!

Signed-off-by: Kipras Melnikovas <kipras@kipras.org>